# docs/swagger.yaml
basePath: /
definitions:
  handlers.apiError:
    properties:
      details:
        type: string
      message:
        type: string
    type: object
  models.Subscription:
    properties:
      end_date:
        example: "2027-05-02T16:21:19Z"
        type: string
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        example: "2026-05-02T16:21:19Z"
        type: string
      user_id:
        type: string
    type: object
  models.SubscriptionCreateReq:
    properties:
      end_date:
        example: 10-2024
        type: string
      price:
        example: 1200
        type: integer
      service_name:
        example: YandexMusic
        type: string
      start_date:
        example: 10-2023
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  models.SubscriptionUpdateReq:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
    type: object
host: localhost:1323
info:
  contact:
    email: lisaosadchenko@gmail.com
  description: This is a documentation for subscription service API
  title: Subscription Service API
  version: "1.0"
paths:
  /subscriptions:
    get:
      consumes:
      - application/json
      description: Get a list of subscriptions with optional filtering and pagination
      parameters:
      - example: 10
        in: query
        maximum: 100
        minimum: -1
        name: limit
        type: integer
      - example: 5
        in: query
        minimum: -1
        name: offset
        type: integer
      - example: 550e8400-e29b-41d4-a716-446655440000
        in: query
        name: userID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched list
          schema:
            items:
              $ref: '#/definitions/models.Subscription'
            type: array
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/handlers.apiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.apiError'
      summary: List subscription records
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Create a subscription record based on provided data
      parameters:
      - description: Subscription details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SubscriptionCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created
          schema:
            additionalProperties:
              $ref: '#/definitions/models.Subscription'
            type: object
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/handlers.apiError'
        "409":
          description: Subscription already exists
          schema:
            $ref: '#/definitions/handlers.apiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.apiError'
      summary: Create subscription record
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a single subscription record by its ID
      parameters:
      - description: Subscription ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted
        "400":
          description: Invalid parameter
          schema:
            $ref: '#/definitions/handlers.apiError'
        "404":
          description: This subscription doesn't exist
          schema:
            $ref: '#/definitions/handlers.apiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.apiError'
      summary: Delete subscription record
      tags:
      - subscriptions
    get:
      consumes:
      - application/json
      description: Get a single subscription record by its ID
      parameters:
      - description: Subscription ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched
          schema:
            $ref: '#/definitions/models.Subscription'
        "400":
          description: Invalid parameter
          schema:
            $ref: '#/definitions/handlers.apiError'
        "404":
          description: This subscription doesn't exist
          schema:
            $ref: '#/definitions/handlers.apiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.apiError'
      summary: Get subscription record
      tags:
      - subscriptions
    patch:
      consumes:
      - application/json
      description: Update an existing subscription by its ID using partial data
      parameters:
      - description: Subscription ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SubscriptionUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated
          schema:
            $ref: '#/definitions/models.Subscription'
        "400":
          description: Invalid parameters / Nothing to update
          schema:
            $ref: '#/definitions/handlers.apiError'
        "404":
          description: This subscription doesn't exist
          schema:
            $ref: '#/definitions/handlers.apiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.apiError'
      summary: Update subscription record
      tags:
      - subscriptions
  /subscriptions/sum:
    get:
      consumes:
      - application/json
      description: Get the sum of prices for subscriptions based on filters
      parameters:
      - example: 10-2024
        in: query
        name: endDate
        type: string
      - example: YandexMusic
        in: query
        name: serviceName
        type: string
      - example: 10-2023
        in: query
        name: startDate
        type: string
      - example: 550e8400-e29b-41d4-a716-446655440000
        in: query
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully calculated total sum
          schema:
            additionalProperties:
              type: number
            type: object
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/handlers.apiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.apiError'
      summary: Calculate total subscriptions price
      tags:
      - subscriptions
swagger: "2.0"